many minor changes to address stricter compiler warnings
git-svn-id: https://oceanai.mit.edu/svn/moos-ivp-aro/trunk@6005 7b6de600-d9ab-45c9-a1ca-e8145d3c6daf

+++ ivp/src/lib_marineview/Shape_Kelp.h
@@ -53,7 +53,7 @@ double g_kelpBody[]=
 
 double       g_kelpCtrX     = 0.0;
 double       g_kelpCtrY     = 0.0;
-unsigned int g_kelpWidth    = 19.62;
+double       g_kelpWidth    = 19.62;
 unsigned int g_kelpBodySize = 24;
 
 #endif
